requirements.txt
Flask==2.1.0
Flask-Login==0.6.2
google-cloud-storage==2.7.0
pytest==6.2.3
pytest-cov==2.11.1
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
Werkzeug==2.2.2
urlparse